groc: Upgrading to v0.6.4 (my branch); on jade 1.x
Because the primary groc project refuses to upgrade to jade 1.x, I did it. It's funny, too, because there have been recent pull requests that make it compatible with jade 1.x, but the package.json trigger has not been pulled.

The good news is that changing that dependency didn't even change the generated html! :0)